+++ ncvue/ncvcontour.py
@@ -45,7 +45,7 @@ from .ncvmethods import set_dim_x, set_dim_y, set_dim_z
 from .ncvwidgets import add_checkbutton, add_combobox, add_entry, add_imagemenu
 from .ncvwidgets import add_spinbox, add_tooltip
 import matplotlib
-matplotlib.use('TkAgg')
+# matplotlib.use('TkAgg')
 from matplotlib import pyplot as plt
 # plt.style.use('seaborn-darkgrid')
 plt.style.use('seaborn-dark')